Module staking::contract

Functions

claim_rewards

Sends available claims to info.sender or as specified in send_to. If claim_as is passed, the claims will be sent as said asset. If restake is true, the accrued ownership will be restaked.

execute

get_deposit_claimables

Get deposit’s claimable fee assets based on which FeeEvents it experienced

instantiate

query

stake

Stake MBRN

unstake

First call is an unstake 2nd call after unstake period is a withdrawal