Module staking::pallet

The module that hosts all the FRAME types needed to add this pallet to a runtime.

Structs

Pallet
The pallet implementing the on-chain logic.

_GeneratedPrefixForStorageNonce

_GeneratedPrefixForStorageRewardPerToken

_GeneratedPrefixForStorageRewardTally

_GeneratedPrefixForStorageSlashPerToken

_GeneratedPrefixForStorageSlashTally

_GeneratedPrefixForStorageStake

_GeneratedPrefixForStorageTotalCurrentStake

_GeneratedPrefixForStorageTotalRewards

_GeneratedPrefixForStorageTotalStake

Enums

Call
Contains one variant per dispatchable that can be called by an extrinsic.

Error
Custom dispatch errors of this pallet.

Event
The event emitted by this pallet.

Traits

Config
Configuration

Type Definitions

Module	Deprecated
Type alias to Pallet, to be used by construct_runtime.

Nonce
The nonce of the current staking pool, used in force refunds. This is a strictly increasing value.

RewardPerToken
Used to compute the rewards for a participant’s stake.

RewardTally
Accounts for previous changes in stake size.

SlashPerToken
Used to compute the amount to slash from a participant’s stake.

SlashTally
Accounts for previous changes in stake size.

Stake
The stake of a participant in this reward pool.

TotalCurrentStake
The total stake - this will increase on deposit and decrease on withdrawal or slashing.

TotalRewards
The total unclaimed rewards distributed to this reward pool. NOTE: this is currently only used for integration tests.

TotalStake
The total stake - this will increase on deposit and decrease on withdrawal.