Module staking::pallet::storage_types

Auto-generated docs-only module listing all (public and private) defined storage types for this pallet.

Warning: Doc-Only

Members of this module cannot be used directly and are only provided for documentation purposes.

To see the actual storage type, find a struct with the same name at the root of the pallet, in the list of Type Definitions.

Structs

Nonce

The nonce of the current staking pool, used in force refunds. This is a strictly increasing value.

RewardPerToken

Used to compute the rewards for a participant’s stake.

RewardTally

Accounts for previous changes in stake size.

SlashPerToken

Used to compute the amount to slash from a participant’s stake.

SlashTally

Accounts for previous changes in stake size.

Stake

The stake of a participant in this reward pool.

TotalCurrentStake

The total stake - this will increase on deposit and decrease on withdrawal or slashing.

TotalRewards

The total unclaimed rewards distributed to this reward pool. NOTE: this is currently only used for integration tests.

TotalStake

The total stake - this will increase on deposit and decrease on withdrawal.