Trait staking::pallet::Config

pub trait Config: Config {
    type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>;
    type SignedInner: CheckedDiv + Ord + FixedPointOperand;
    type SignedFixedPoint: FixedPointNumber<Inner = Self::SignedInner> + TruncateFixedPointToInt + Encode + EncodeLike + Decode + TypeInfo + MaxEncodedLen;
    type GetNativeCurrencyId: Get<Self::CurrencyId>;
    type CurrencyId: Parameter + Member + Copy + MaybeSerializeDeserialize + Ord + MaxEncodedLen;
}

Configuration trait of this pallet.

The main purpose of this trait is to act as an interface between this pallet and the runtime in which it is embedded in. A type, function, or constant in this trait is essentially left to be configured by the runtime that includes this pallet.

Consequently, a runtime that wants to include this pallet must implement this trait.

Configuration

The pallet’s configuration trait.

Required Associated Types

type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>

The overarching event type.

type SignedInner: CheckedDiv + Ord + FixedPointOperand

The Inner type of the SignedFixedPoint.

type SignedFixedPoint: FixedPointNumber<Inner = Self::SignedInner> + TruncateFixedPointToInt + Encode + EncodeLike + Decode + TypeInfo + MaxEncodedLen

Signed fixed point type.

type GetNativeCurrencyId: Get<Self::CurrencyId>

type CurrencyId: Parameter + Member + Copy + MaybeSerializeDeserialize + Ord + MaxEncodedLen

The currency ID type.